Remove pythia dep; migrate CloudKeyManager to BrainkeyClient (foundation-android) (#72)
* Remove pythia dep; migrate CloudKeyManager to BrainkeyClient from foundation-android

- Remove com.virgilsecurity:pythia-android and com.virgilsecurity.crypto:pythia-android
- Add com.virgilsecurity.crypto:foundation-android:0.19.0-rc.11 explicitly
- Bump virgilSdk 7.4.0→7.5.0, ratchet 0.3.0→0.4.0, virgilCrypto 0.17.2→0.19.0-rc.11
- Remove sunset jcenter repository
- New BrainkeyHttpClient: POST /pythia/v1/brainkey with BrainkeyClient OPRF protocol
  (ed25519 key derivation; key name "e3kit-backup" as HKDF domain separator)
- CloudKeyManager: replace BrainKey/BrainKeyContext/VirgilPythiaCrypto with deriveKeyPair()
  delegating to BrainkeyHttpClient.deriveKeyPair(password, crypto)
- Update instrumentation tests to use BrainkeyHttpClient instead of pythia SDK

* Fix Int-to-Long widening for GroupSessionMessage.epoch in Group.kt

virgilSdk 7.5.0 changed GroupManager.retrieve(sessionId, epoch) to expect
Long; add explicit .toLong() cast to satisfy the call site.

+++ ethree-common/src/main/java/com/virgilsecurity/android/common/model/Group.kt
@@ -195,7 +195,7 @@ class Group internal constructor(
             } else {
                 val sessionId = encrypted.sessionId.toData()
 
-                val tempGroup = this.groupManager.retrieve(sessionId, messageEpoch)
+                val tempGroup = this.groupManager.retrieve(sessionId, messageEpoch.toLong())
                                 ?: throw GroupException(
                                     GroupException.Description.MISSING_CACHED_GROUP
                                 )
